create new kong2kic package and move code
