+++ kong2kic/testdata/kong2kic/annotations/json/expected-output.json
