+++ kong2kic/testdata/kong2kic/annotations/yaml/expected-output.yaml
