+++ kong2kic/testdata/kong2kic/custom_resources/json/expected-output.json
