+++ kong2kic/testdata/kong2kic/custom_resources/yaml/expected-output.yaml
