+++ kong2kic/testdata/kong2kic/gateway/json/expected-output.json
