+++ kong2kic/testdata/kong2kic/gateway/yaml/expected-output.yaml
