+++ kong2kic/testdata/kong2kic/input.yaml
